# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  company?: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface BlogPublishInput {
  /**
     * @minLength 8
     * @maxLength 160
     */
  title: string;
  /**
     * @minLength 2
     * @maxLength 40
     */
  category: string;
  /**
     * @minLength 100
     * @maxLength 100000
     */
  articleText: string;
  /**
     * Cover image as a base64 data URL (png, jpeg or webp)
     * @minLength 50
     * @maxLength 16000000
     */
  imageDataUrl: string;
  /** @maxLength 300 */
  excerpt?: string;
  /** @maxLength 80 */
  author?: string;
}

export interface BlogPublishResponse {
  ok: boolean;
  slug: string;
  url: string;
}

export type GalleryCategory = typeof GalleryCategory[keyof typeof GalleryCategory];


export const GalleryCategory = {
  residential: 'residential',
  commercial: 'commercial',
  industrial: 'industrial',
  solar: 'solar',
  battery: 'battery',
  ev: 'ev',
  'hybrid-inverter': 'hybrid-inverter',
  lifepo4: 'lifepo4',
} as const;

export interface GalleryPublishInput {
  category: GalleryCategory;
  /**
     * @minItems 1
     * @maxItems 20
     * @items.minLength 50
     * @items.maxLength 16000000
     */
  images: string[];
}

export interface GalleryPublishResponse {
  ok: boolean;
  count: number;
}