# lib/api-zod/src/generated/types/galleryCategory.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type GalleryCategory = typeof GalleryCategory[keyof typeof GalleryCategory];


export const GalleryCategory = {
  residential: 'residential',
  commercial: 'commercial',
  industrial: 'industrial',
  solar: 'solar',
  battery: 'battery',
  ev: 'ev',
  'hybrid-inverter': 'hybrid-inverter',
  lifepo4: 'lifepo4',
} as const;