# lib/api-zod/src/generated/types/galleryPublishInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GalleryCategory } from './galleryCategory';

export interface GalleryPublishInput {
  category: GalleryCategory;
  /**
     * @minItems 1
     * @maxItems 20
     * @items.minLength 50
     * @items.maxLength 16000000
     */
  images: string[];
}