# lib/api-zod/src/generated/types/galleryPublishResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryPublishResponse {
  ok: boolean;
  count: number;
}